Minor Changes, bugfixes. Start US 13
- Made some seperation of converns related changes:
	- Moved Geometry related calculations (i.e. Bresenham) into
	static GeometryCalculator Class.
- Streamlined drawing and canvas code in Form1, by moving functionality
into helping functions.
- Fixed a bug where drawing, making the window smaller and then deleting
a line would update the pixeldensity of the canvas.
- Fixed unreliable deleting caused by the curserPositions queue always
being filled and emptied immediatly.
- Moved userstory.md files around, updated them
- Prepared for starting Userstory 13
