+++ SketchAssistant/SketchAssistant/SketchAssistant.csproj
@@ -71,6 +71,7 @@
   <ItemGroup>
     <Compile Include="FileImporter.cs" />
     <Compile Include="FileImporterException.cs" />
+    <Compile Include="GeometryCalculator.cs" />
     <Compile Include="SketchAction.cs" />
     <Compile Include="ActionHistory.cs" />
     <Compile Include="Line.cs" />
